feat(pack-code): Add password file check and read password for packing process

+++ pack-code.sh
@@ -2,6 +2,8 @@
 
 # cd ./mongodb-csfle-migrator
 
+PASSWORD_FILE_PATH="password.hidden.txt"
+
 PACK_CODE_FOLDER="./tmp/packed"
 mkdir -p $PACK_CODE_FOLDER
 PACK_CODE_FOLDER=$(realpath "$PACK_CODE_FOLDER")
@@ -18,6 +20,14 @@ TARGET_SAMPLE_DATA_FOLDER="$PACK_CODE_FOLDER/.tmp/sample-data"
 
 TARGET_OUTPUT_FOLDER="$PACK_CODE_FOLDER/OUTPUT"
 
+# Check if file exists
+if [[ ! -f "$PASSWORD_FILE_PATH" ]]; then
+  echo "Error: File '$PASSWORD_FILE_PATH' not found."
+  exit 1
+fi
+
+# Read value from file (first line, as example)
+PASSWORD=$(<"$PASSWORD_FILE_PATH")
 
 # Clean up the target folders
 rm -rf $PACK_CODE_FOLDER/*
@@ -51,9 +61,6 @@ cp -r $SOURCE_SAMPLE_DATA_FOLDER/* $TARGET_SAMPLE_DATA_FOLDER
 # Copy the docs to the target folder
 cp -r $SOURCE_DOC_FOLDER/* $TARGET_DOC_FOLDER
 
-
-PASSWORD=NGQMGY6V
-
 echo "Packing code..."
 echo "Packing code to $TARGET_OUTPUT_FOLDER"
 echo "Packing code to $TARGET_OUTPUT_FOLDER/.tmp"
@@ -72,3 +79,5 @@ TARGET_OUTPUT_TMP_FOLDER="$TARGET_OUTPUT_FOLDER/.tmp"
 $TARGET_OUTPUT_TMP_FOLDER/sample-test-data.zip \
 $TARGET_OUTPUT_TMP_FOLDER/source_code.zip \
 $TARGET_OUTPUT_TMP_FOLDER/docs.zip
+
+echo "Packing with Password: $PASSWORD"
